Fixing some bug!🛠️
So I spent my day finding the bug on why for some reason the @media queries doesn't work and also my JS giving an error that said "Image is null"

Upon searching I finally find what is the problem. I didn't put:

  document.addEventListener("DOMContentLoaded", function ()

...to my JS.

with this piece of code, finally my @media-queries are working.

Those background-color that;s in the CSS are only for testing purposes and it will be remove once I've manage to style some more using @media queries